+++ tests/features/monrun.feature
@@ -147,10 +147,7 @@ Feature: ch-monitoring tool
     """
     INSERT INTO test.table_01 SELECT number FROM numbers(100)
     """
-    And we execute command on clickhouse02
-    """
-    sleep 5
-    """
+    And we sleep for 5 seconds
     And we execute command on clickhouse01
     """
     ch-monitoring replication-lag -w 4
